requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.37
pydantic-settings==2.7.1
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
celery==5.4.0
redis==5.2.1
pytest==8.3.4
httpx==0.28.1
psycopg[binary]==3.2.9
pypdf==5.4.0
pgvector==0.3.6
sentence-transformers==3.4.1
email-validator==2.3.0
